Add ChatInterface and STMEditor components for enhanced user interaction
- Introduced ChatInterface component to manage chat functionality, including tool call handling and message display.
- Created STMEditor component for managing STM state, allowing users to add, edit, and delete STM keys with associated attributes.
- Updated ClientSTMDemo to integrate both components, streamlining the user experience and improving state management.
- Implemented keyboard shortcuts for STM operations, enhancing usability and efficiency.
